feat: BSZ-FRUN-01-runtime-session-surfaces-004 - Preserve Petri guard semantics coverage under file size limit
