+++ pkg/services/factory_runtime/internal/orchestrators/petri/guard_test.go
@@ -624,355 +624,6 @@ func TestAnyWithParentGuard_MissingBinding(t *testing.T) {
 	}
 }
 
-func TestDependencyGuard_AllDependenciesMet(t *testing.T) {
-	// Dependency token A is in the required "complete" state.
-	depToken := &factorytoken.Token{
-		ID:      "tok-a",
-		PlaceID: "task:complete",
-		Color: factorytoken.Color{
-			WorkID:     "work-a",
-			WorkTypeID: "task",
-		},
-	}
-
-	// Candidate B depends on A being in "complete".
-	candidates := []factorytoken.Token{
-		{
-			ID:      "tok-b",
-			PlaceID: "task:init",
-			Color: factorytoken.Color{
-				WorkID:     "work-b",
-				WorkTypeID: "task",
-				Relations: []work.Relation{
-					{Type: work.RelationDependsOn, TargetWorkID: "work-a", RequiredState: "complete"},
-				},
-			},
-		},
-	}
-
-	marking := &MarkingSnapshot{
-		Tokens: map[string]*factorytoken.Token{
-			"tok-a": depToken,
-			"tok-b": &candidates[0],
-		},
-	}
-
-	guard := &DependencyGuard{}
-	matched, ok := guard.Evaluate(candidates, nil, marking)
-	if !ok {
-		t.Fatal("expected guard to pass when dependency is in required state")
-	}
-	if len(matched) != 1 || matched[0].ID != "tok-b" {
-		t.Errorf("expected tok-b matched, got %v", matched)
-	}
-}
-
-func TestDependencyGuard_DependencyNotMet(t *testing.T) {
-	// Dependency token A is in "init" — not in "complete".
-	depToken := &factorytoken.Token{
-		ID:      "tok-a",
-		PlaceID: "task:init",
-		Color: factorytoken.Color{
-			WorkID:     "work-a",
-			WorkTypeID: "task",
-		},
-	}
-
-	candidates := []factorytoken.Token{
-		{
-			ID:      "tok-b",
-			PlaceID: "task:init",
-			Color: factorytoken.Color{
-				WorkID:     "work-b",
-				WorkTypeID: "task",
-				Relations: []work.Relation{
-					{Type: work.RelationDependsOn, TargetWorkID: "work-a", RequiredState: "complete"},
-				},
-			},
-		},
-	}
-
-	marking := &MarkingSnapshot{
-		Tokens: map[string]*factorytoken.Token{
-			"tok-a": depToken,
-			"tok-b": &candidates[0],
-		},
-	}
-
-	guard := &DependencyGuard{}
-	matched, ok := guard.Evaluate(candidates, nil, marking)
-	if ok {
-		t.Fatal("expected guard to fail when dependency is not in required state")
-	}
-	if len(matched) != 0 {
-		t.Errorf("expected 0 matches, got %d", len(matched))
-	}
-}
-
-func TestDependencyGuard_DependencyNotFound(t *testing.T) {
-	candidates := []factorytoken.Token{
-		{
-			ID:      "tok-b",
-			PlaceID: "task:init",
-			Color: factorytoken.Color{
-				WorkID:     "work-b",
-				WorkTypeID: "task",
-				Relations: []work.Relation{
-					{Type: work.RelationDependsOn, TargetWorkID: "work-missing", RequiredState: "complete"},
-				},
-			},
-		},
-	}
-
-	marking := &MarkingSnapshot{
-		Tokens: map[string]*factorytoken.Token{
-			"tok-b": &candidates[0],
-		},
-	}
-
-	guard := &DependencyGuard{}
-	matched, ok := guard.Evaluate(candidates, nil, marking)
-	if ok {
-		t.Fatal("expected guard to fail when dependency token is missing")
-	}
-	if len(matched) != 0 {
-		t.Errorf("expected 0 matches, got %d", len(matched))
-	}
-}
-
-func TestDependencyGuard_NilMarking(t *testing.T) {
-	candidates := []factorytoken.Token{
-		{
-			ID: "tok-b",
-			Color: factorytoken.Color{
-				Relations: []work.Relation{
-					{Type: work.RelationDependsOn, TargetWorkID: "work-a", RequiredState: "complete"},
-				},
-			},
-		},
-	}
-
-	guard := &DependencyGuard{}
-	matched, ok := guard.Evaluate(candidates, nil, nil)
-	if ok {
-		t.Fatal("expected guard to fail with nil marking")
-	}
-	if matched != nil {
-		t.Errorf("expected nil matches, got %v", matched)
-	}
-}
-
-func TestDependencyGuard_NoDependencies(t *testing.T) {
-	// Token with no DEPENDS_ON relations should pass.
-	candidates := []factorytoken.Token{
-		{
-			ID:      "tok-b",
-			PlaceID: "task:init",
-			Color: factorytoken.Color{
-				WorkID:     "work-b",
-				WorkTypeID: "task",
-				Relations: []work.Relation{
-					{Type: work.RelationParentChild, TargetWorkID: "work-a"},
-				},
-			},
-		},
-	}
-
-	marking := &MarkingSnapshot{
-		Tokens: map[string]*factorytoken.Token{
-			"tok-b": &candidates[0],
-		},
-	}
-
-	guard := &DependencyGuard{}
-	matched, ok := guard.Evaluate(candidates, nil, marking)
-	if !ok {
-		t.Fatal("expected guard to pass for token with no DEPENDS_ON relations")
-	}
-	if len(matched) != 1 {
-		t.Errorf("expected 1 match, got %d", len(matched))
-	}
-}
-
-func TestDependencyGuard_MultipleDependencies(t *testing.T) {
-	depA := &factorytoken.Token{
-		ID:      "tok-a",
-		PlaceID: "task:complete",
-		Color:   factorytoken.Color{WorkID: "work-a", WorkTypeID: "task"},
-	}
-	depC := &factorytoken.Token{
-		ID:      "tok-c",
-		PlaceID: "task:complete",
-		Color:   factorytoken.Color{WorkID: "work-c", WorkTypeID: "task"},
-	}
-
-	candidates := []factorytoken.Token{
-		{
-			ID:      "tok-b",
-			PlaceID: "task:init",
-			Color: factorytoken.Color{
-				WorkID:     "work-b",
-				WorkTypeID: "task",
-				Relations: []work.Relation{
-					{Type: work.RelationDependsOn, TargetWorkID: "work-a", RequiredState: "complete"},
-					{Type: work.RelationDependsOn, TargetWorkID: "work-c", RequiredState: "complete"},
-				},
-			},
-		},
-	}
-
-	marking := &MarkingSnapshot{
-		Tokens: map[string]*factorytoken.Token{
-			"tok-a": depA,
-			"tok-b": &candidates[0],
-			"tok-c": depC,
-		},
-	}
-
-	guard := &DependencyGuard{}
-	matched, ok := guard.Evaluate(candidates, nil, marking)
-	if !ok {
-		t.Fatal("expected guard to pass when all dependencies are met")
-	}
-	if len(matched) != 1 {
-		t.Errorf("expected 1 match, got %d", len(matched))
-	}
-}
-
-func TestDependencyGuard_PartialDependenciesMet(t *testing.T) {
-	depA := &factorytoken.Token{
-		ID:      "tok-a",
-		PlaceID: "task:complete",
-		Color:   factorytoken.Color{WorkID: "work-a", WorkTypeID: "task"},
-	}
-	depC := &factorytoken.Token{
-		ID:      "tok-c",
-		PlaceID: "task:init", // NOT complete
-		Color:   factorytoken.Color{WorkID: "work-c", WorkTypeID: "task"},
-	}
-
-	candidates := []factorytoken.Token{
-		{
-			ID:      "tok-b",
-			PlaceID: "task:init",
-			Color: factorytoken.Color{
-				WorkID:     "work-b",
-				WorkTypeID: "task",
-				Relations: []work.Relation{
-					{Type: work.RelationDependsOn, TargetWorkID: "work-a", RequiredState: "complete"},
-					{Type: work.RelationDependsOn, TargetWorkID: "work-c", RequiredState: "complete"},
-				},
-			},
-		},
-	}
-
-	marking := &MarkingSnapshot{
-		Tokens: map[string]*factorytoken.Token{
-			"tok-a": depA,
-			"tok-b": &candidates[0],
-			"tok-c": depC,
-		},
-	}
-
-	guard := &DependencyGuard{}
-	matched, ok := guard.Evaluate(candidates, nil, marking)
-	if ok {
-		t.Fatal("expected guard to fail when only some dependencies are met")
-	}
-	if len(matched) != 0 {
-		t.Errorf("expected 0 matches, got %d", len(matched))
-	}
-}
-
-func TestFanoutCountGuard_ExactMatch(t *testing.T) {
-	parent := &factorytoken.Token{Color: factorytoken.Color{WorkID: "parent-1"}}
-	countToken := &factorytoken.Token{Color: factorytoken.Color{Tags: map[string]string{"expected_count": "3"}}}
-	bindings := map[string]*factorytoken.Token{"parent": parent, "fanout-count": countToken}
-
-	candidates := []factorytoken.Token{
-		{ID: "c1", Color: factorytoken.Color{ParentID: "parent-1"}},
-		{ID: "c2", Color: factorytoken.Color{ParentID: "parent-1"}},
-		{ID: "c3", Color: factorytoken.Color{ParentID: "parent-1"}},
-		{ID: "c4", Color: factorytoken.Color{ParentID: "other"}},
-	}
-
-	guard := &FanoutCountGuard{MatchBinding: "parent", CountBinding: "fanout-count"}
-	matched, ok := guard.Evaluate(candidates, bindings, nil)
-	if !ok {
-		t.Fatal("expected guard to pass with 3 matching children and expected_count=3")
-	}
-	if len(matched) != 3 {
-		t.Fatalf("expected 3 matches, got %d", len(matched))
-	}
-}
-
-func TestFanoutCountGuard_CountMismatch(t *testing.T) {
-	parent := &factorytoken.Token{Color: factorytoken.Color{WorkID: "parent-1"}}
-	countToken := &factorytoken.Token{Color: factorytoken.Color{Tags: map[string]string{"expected_count": "3"}}}
-	bindings := map[string]*factorytoken.Token{"parent": parent, "fanout-count": countToken}
-
-	// Only 2 children — expected 3.
-	candidates := []factorytoken.Token{
-		{ID: "c1", Color: factorytoken.Color{ParentID: "parent-1"}},
-		{ID: "c2", Color: factorytoken.Color{ParentID: "parent-1"}},
-	}
-
-	guard := &FanoutCountGuard{MatchBinding: "parent", CountBinding: "fanout-count"}
-	_, ok := guard.Evaluate(candidates, bindings, nil)
-	if ok {
-		t.Fatal("expected guard to fail when count doesn't match")
-	}
-}
-
-func TestFanoutCountGuard_ZeroChildren(t *testing.T) {
-	parent := &factorytoken.Token{Color: factorytoken.Color{WorkID: "parent-1"}}
-	countToken := &factorytoken.Token{Color: factorytoken.Color{Tags: map[string]string{"expected_count": "0"}}}
-	bindings := map[string]*factorytoken.Token{"parent": parent, "fanout-count": countToken}
-
-	guard := &FanoutCountGuard{MatchBinding: "parent", CountBinding: "fanout-count"}
-	matched, ok := guard.Evaluate(nil, bindings, nil)
-	if !ok {
-		t.Fatal("expected guard to pass with 0 children and expected_count=0")
-	}
-	if len(matched) != 0 {
-		t.Fatalf("expected 0 matches, got %d", len(matched))
-	}
-}
-
-func TestFanoutCountGuard_MissingParentBinding(t *testing.T) {
-	countToken := &factorytoken.Token{Color: factorytoken.Color{Tags: map[string]string{"expected_count": "1"}}}
-	bindings := map[string]*factorytoken.Token{"fanout-count": countToken}
-
-	guard := &FanoutCountGuard{MatchBinding: "parent", CountBinding: "fanout-count"}
-	_, ok := guard.Evaluate(nil, bindings, nil)
-	if ok {
-		t.Fatal("expected guard to fail when parent binding is missing")
-	}
-}
-
-func TestFanoutCountGuard_MissingCountBinding(t *testing.T) {
-	parent := &factorytoken.Token{Color: factorytoken.Color{WorkID: "parent-1"}}
-	bindings := map[string]*factorytoken.Token{"parent": parent}
-
-	guard := &FanoutCountGuard{MatchBinding: "parent", CountBinding: "fanout-count"}
-	_, ok := guard.Evaluate(nil, bindings, nil)
-	if ok {
-		t.Fatal("expected guard to fail when count binding is missing")
-	}
-}
-
-func TestFanoutCountGuard_InvalidCountTag(t *testing.T) {
-	parent := &factorytoken.Token{Color: factorytoken.Color{WorkID: "parent-1"}}
-	countToken := &factorytoken.Token{Color: factorytoken.Color{Tags: map[string]string{"expected_count": "not-a-number"}}}
-	bindings := map[string]*factorytoken.Token{"parent": parent, "fanout-count": countToken}
-
-	guard := &FanoutCountGuard{MatchBinding: "parent", CountBinding: "fanout-count"}
-	_, ok := guard.Evaluate(nil, bindings, nil)
-	if ok {
-		t.Fatal("expected guard to fail with invalid expected_count")
-	}
-}
-
 func TestTokenColorField(t *testing.T) {
 	color := factorytoken.Color{
 		WorkID:     "w-1",
